Move unhandled node attributes gate to `Node` class
